HIR-12 rename : git 충돌 방지를 위해 디렉토리 구조 수정
