--- src/main/java/kr/binarybard/hireo/auth/service/LoginService.java
@@ -1,29 +1,0 @@
-package kr.binarybard.hireo.auth.service;
-
-import kr.binarybard.hireo.exception.MemberNotFoundException;
-import kr.binarybard.hireo.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class LoginService implements UserDetailsService {
-	private final MemberRepository memberRepository;
-
-	@Override
-	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		var member = memberRepository.findByEmail(username)
-			.orElseThrow(MemberNotFoundException::new);
-		return User.builder()
-			.username(member.getEmail())
-			.password(member.getPassword())
-			.build();
-	}
-}
