--- src/main/java/kr/binarybard/hireo/company/repository/CompanyRepository.java
@@ -1,8 +1,0 @@
-package kr.binarybard.hireo.company.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import kr.binarybard.hireo.company.domain.Company;
-
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-}
