+++ src/main/java/kr/binarybard/hireo/config/oauth/CustomOAuth2UserService.java
@@ -1,11 +1,8 @@
 package kr.binarybard.hireo.config.oauth;
 
-import kr.binarybard.hireo.config.oauth.userinfo.CustomOAuth2UserInfo;
-import kr.binarybard.hireo.member.domain.Member;
-import kr.binarybard.hireo.member.domain.Role;
-import kr.binarybard.hireo.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.util.Optional;
+import java.util.UUID;
+
 import org.springframework.context.annotation.Lazy;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
@@ -14,8 +11,11 @@ import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
-import java.util.UUID;
+import kr.binarybard.hireo.config.oauth.userinfo.CustomOAuth2UserInfo;
+import kr.binarybard.hireo.web.member.domain.Member;
+import kr.binarybard.hireo.web.member.domain.Role;
+import kr.binarybard.hireo.web.member.repository.MemberRepository;
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
