--- src/main/java/kr/binarybard/hireo/member/controller/MainController.java
@@ -1,13 +1,0 @@
-package kr.binarybard.hireo.member.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-	@GetMapping("/")
-	public String index() {
-		return "index";
-	}
-}
