--- src/main/java/kr/binarybard/hireo/member/domain/Role.java
@@ -1,3 +1,0 @@
-package kr.binarybard.hireo.member.domain;
-
-public enum Role { FREELANCER, EMPLOYER, ADMINISTRATOR }
