+++ src/main/java/kr/binarybard/hireo/web/auth/service/LoginService.java
@@ -1,0 +1,29 @@
+package kr.binarybard.hireo.web.auth.service;
+
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import kr.binarybard.hireo.exception.MemberNotFoundException;
+import kr.binarybard.hireo.web.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class LoginService implements UserDetailsService {
+	private final MemberRepository memberRepository;
+
+	@Override
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		var member = memberRepository.findByEmail(username)
+			.orElseThrow(MemberNotFoundException::new);
+		return User.builder()
+			.username(member.getEmail())
+			.password(member.getPassword())
+			.build();
+	}
+}
