+++ src/main/java/kr/binarybard/hireo/web/company/dto/CompanyMapper.java
@@ -1,0 +1,19 @@
+package kr.binarybard.hireo.web.company.dto;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.factory.Mappers;
+
+import kr.binarybard.hireo.web.company.domain.Company;
+import kr.binarybard.hireo.web.location.dto.LocationMapper;
+
+@Mapper(componentModel = "spring", uses = LocationMapper.class)
+public interface CompanyMapper {
+	CompanyMapper INSTANCE = Mappers.getMapper(CompanyMapper.class);
+
+	@Mapping(target = "location", source = "locationDto")
+	Company toEntity(CompanyRegister companyDto);
+
+	@Mapping(target = "locationDto", source = "location")
+	CompanyResponse toDto(Company company);
+}
