+++ src/main/java/kr/binarybard/hireo/web/member/controller/MainController.java
@@ -1,0 +1,13 @@
+package kr.binarybard.hireo.web.member.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+	@GetMapping("/")
+	public String index() {
+		return "index";
+	}
+}
