+++ src/test/java/kr/binarybard/hireo/member/service/MemberServiceTest.java
@@ -1,21 +1,11 @@
 package kr.binarybard.hireo.member.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.ArgumentMatchers.*;
+import static org.mockito.Mockito.*;
 
 import java.util.Optional;
-import kr.binarybard.hireo.auth.dto.SignUpRequest;
-import kr.binarybard.hireo.exception.MemberNotFoundException;
-import kr.binarybard.hireo.fixture.LoginFixture;
-import kr.binarybard.hireo.fixture.MemberFixture;
-import kr.binarybard.hireo.member.domain.Member;
-import kr.binarybard.hireo.member.dto.MemberMapper;
-import kr.binarybard.hireo.member.repository.MemberRepository;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -23,6 +13,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
+
+import kr.binarybard.hireo.exception.MemberNotFoundException;
+import kr.binarybard.hireo.fixture.LoginFixture;
+import kr.binarybard.hireo.fixture.MemberFixture;
+import kr.binarybard.hireo.web.auth.dto.SignUpRequest;
+import kr.binarybard.hireo.web.member.domain.Member;
+import kr.binarybard.hireo.web.member.dto.MemberMapper;
+import kr.binarybard.hireo.web.member.repository.MemberRepository;
+import kr.binarybard.hireo.web.member.service.MemberService;
 
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
